@@ -1,30 +1,32 @@
 import React from "react";
-import {withStyles, Button, Typography, Grid, Fab, Box} from "@material-ui/core";
-import ArrowBack from "@material-ui/icons/ArrowBack";
-import Business from "@material-ui/icons/Business";
-import Email from "@material-ui/icons/Email";
-import Call from "@material-ui/icons/Call";
-import Chip from "@material-ui/core/Chip";
-
+import {Button, Typography, Grid, Fab, Box, Chip} from "@mui/material";
+import {withStyles} from "@mui/styles"
+import ArrowBack from "@mui/icons-material/ArrowBack";
+import Business from "@mui/icons-material/Business";
+import Email from "@mui/icons-material/Email";
+import Call from "@mui/icons-material/Call";
 
 const styles = theme => ({
-   root: {
-       padding: theme.spacing(1),
-       minHeight: 700
-   },
-    item: {
-       maxWidth: 1000
+    root: {
+        flexGrow: 1,
+        padding: theme.spacing(1),
+        paddingTop: theme.spacing(4),
+        paddingBottom: theme.spacing(4),
+        minHeight: 700
     },
-    button:{
+    /*item: {
+        maxWidth: 1200
+    },*/
+   /* button:{
         marginTop : theme.spacing(1),
         marginBottom : theme.spacing(3),
         background: '#ffe01b'
-    },
-    fab: {
+    },*/
+    /*fab: {
         marginTop: theme.spacing(1),
         marginBottom: theme.spacing(3),
         background: '#ffe01b'
-    },
+    },*/
     screenshot:{
         borderStyle: "solid",
         borderWidth: 1,
@@ -37,140 +39,177 @@ const styles = theme => ({
     icon: {
         marginRight: theme.spacing(2)
     },
-    chip: {
+    /*chip: {
         borderRadius: "3px",
         marginRight: theme.spacing(1)
-    },
+    },*/
     mediaBox: {
         paddingTop: theme.spacing(2),
         maxWidth: 800
     },
+    paper: {
+        backgroundColor: theme.palette.background.opaque,
+        padding: theme.spacing(2),
+        color: theme.palette.primario.contrastText,
+        borderStyle: 'solid',
+        borderWidth: 1,
+        borderColor: theme.palette.secundario.main,
+        borderRadius: '10px 10px 10px 10px',
+        display: 'flex',
+        justifyContent: "center"
+    },
+    box: {
+        maxWidth: '900px', paddingTop: '50px', paddingBottom: '50px'
+    }
 });
 
 const ProductDetails = props => {
     const {classes, product, setSelected} = props;
 
     return <div className={classes.root}>
-        <Grid container justify="center">
-            <Grid item xs={12} className={classes.item}>
-                <Fab onClick={ () => setSelected(null)} className={classes.fab} size="medium">
-                    <ArrowBack/>
-                </Fab>
+        <Grid container justifyContent="center">
+            <Grid item xs={12} style={{maxWidth: 1200}}>
 
-                <Typography variant="h5">
-                    {product.name}
-                </Typography>
-                <Typography variant="subtitle1" color="textSecondary" paragraph>
-                    Por {product.maintainer}
-                </Typography>
+                <div className={classes.paper}>
+                    <Box className={classes.box}>
 
-                <Typography paragraph align="justify">
-                    {product.description}
-                </Typography>
+                        <Fab onClick={ () => setSelected(null)} size="medium" style={{
+                            marginTop: 5,
+                            marginBottom: 20,
+                            background: '#ffe01b'
+                        }}>
+                            <ArrowBack/>
+                        </Fab>
 
-                <Button variant="contained" href={product.url} target="_blank" className={classes.button}>
-                    Conoce más
-                </Button>
-
-                <Typography variant="h6">
-                    Contacto
-                </Typography>
-
-                {product.contactPoint.directory_url?
-                    <Box paddingTop={1}>
-                        <Typography paragraph align="justify">
-                            Si tienes dudas sobre la instalación del Sistema de Declaración Patrimonial y de Intereses
-                            de la PDN comunicate con alguno de los miembros de la Red Nacional de Capacitadores (RNC).
+                        <Typography variant="h5">
+                            {product.name}
+                        </Typography>
+                        <Typography variant="subtitle1" paragraph>
+                            Por {product.maintainer}
                         </Typography>
 
-                        <Button variant="contained"
-                                href={product.contactPoint.directory_url}
-                                target="_blank" className={classes.button}>
-                            Directorio RNC
+                        <Typography paragraph align="justify">
+                            {product.description}
+                        </Typography>
+
+                        <Button variant="contained" href={product.url} target="_blank" color="secundario" sx={{
+                            marginTop : 1,
+                            marginBottom : 3,
+                        }}>
+                            Conoce más
                         </Button>
-                    </Box>
-                    :
-                    <div>
-                        <Box display="flex" flexWrap="wrap" p={1}>
-                            <Business className={classes.icon}/>
-                            <Typography>
-                                {product.contactPoint.organization}
-                            </Typography>
-                        </Box>
 
-                        <Box display="flex" flexWrap="wrap" p={1}>
-                            <Call className={classes.icon}/>
-                            <Typography>
-                                {product.contactPoint.phone}
-                            </Typography>
-                        </Box>
+                        <Typography variant="h6">
+                            Contacto
+                        </Typography>
 
-                        <Box display="flex" flexWrap="wrap" p={1} paddingBottom={3}>
-                            <Email className={classes.icon}/>
-                            <Typography>
-                                {product.contactPoint.email}
-                            </Typography>
-                        </Box>
-                    </div>
-                }
+                        {product.contactPoint.directory_url?
+                            <Box paddingTop={1}>
+                                <Typography paragraph align="justify">
+                                    Si tienes dudas sobre la instalación del Sistema de Declaración Patrimonial y de Intereses
+                                    de la PDN comunicate con alguno de los miembros de la Red Nacional de Capacitadores (RNC).
+                                </Typography>
 
-                <Typography variant="h6" paragraph>
-                    Tecnologías
-                </Typography>
-
-                <Box paddingBottom={2}>
-                    {
-                        product.technologies.map((t,i) => <Chip
-                            key={i}
-                            label={t}
-                            size="small"
-                            className={classes.chip}
-                        />)
-                    }
-                </Box>
-
-                {product.media && Array.isArray(product.media) && product.media.length > 0 &&
-                <div>
-                    <Box display="flex" flexWrap="wrap">
-                        {
-                            product.media.map((s, i) => {
-                                return <Box key={i} p={1} className={classes.mediaBox}>
-                                    <Typography variant="h6">
-                                        {product.media[i].title}
+                                <Button variant="contained"
+                                        href={product.contactPoint.directory_url}
+                                        target="_blank"  color="secundario" sx={{
+                                        marginTop : 1,
+                                        marginBottom : 3,
+                                }}>
+                                    Directorio RNC
+                                </Button>
+                            </Box>
+                            :
+                            <div>
+                                <Box display="flex" flexWrap="wrap" p={1}>
+                                    <Business className={classes.icon}/>
+                                    <Typography>
+                                        {product.contactPoint.organization}
                                     </Typography>
-
-                                    <iframe width="100%" height="450" src={product.media[i].url}
-                                            title="YouTube video player" frameBorder="0"
-                                            allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture"
-                                            allowFullScreen/>
-
-                                    <Typography variant="body2" paragraph align="justify">
-                                        {product.media[i].caption}
-                                    </Typography>
-
                                 </Box>
-                            })
+
+                                <Box display="flex" flexWrap="wrap" p={1}>
+                                    <Call className={classes.icon}/>
+                                    <Typography>
+                                        {product.contactPoint.phone}
+                                    </Typography>
+                                </Box>
+
+                                <Box display="flex" flexWrap="wrap" p={1} paddingBottom={3}>
+                                    <Email className={classes.icon}/>
+                                    <Typography>
+                                        {product.contactPoint.email}
+                                    </Typography>
+                                </Box>
+                            </div>
                         }
+
+                        <Typography variant="h6" paragraph>
+                            Tecnologías
+                        </Typography>
+
+                        <Box paddingBottom={2}>
+                            {
+                                product.technologies.map((t,i) => <Chip
+                                    key={i}
+                                    label={t}
+                                    size="small"
+                                    style={{
+                                        color: "#ffffff",
+                                        fontWeight: "bold",
+                                        borderRadius: "3px",
+                                        marginRight: 5
+                                    }}
+                                />)
+                            }
+                        </Box>
+
+                        {product.media && Array.isArray(product.media) && product.media.length > 0 &&
+                            <div>
+                                <Box display="flex" flexWrap="wrap">
+                                    {
+                                        product.media.map((s, i) => {
+                                            return <Box key={i} p={1} className={classes.mediaBox}>
+                                                <Typography variant="h6">
+                                                    {product.media[i].title}
+                                                </Typography>
+
+                                                <iframe width="100%" height="450" src={product.media[i].url}
+                                                        title="YouTube video player" frameBorder="0"
+                                                        allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture"
+                                                        allowFullScreen/>
+
+                                                <Typography variant="body2" paragraph align="justify">
+                                                    {product.media[i].caption}
+                                                </Typography>
+
+                                            </Box>
+                                        })
+                                    }
+                                </Box>
+                            </div>
+                        }
+
+                        {product.screenshots && Array.isArray(product.screenshots) && product.screenshots.length > 0 &&
+                            <div>
+                                <Typography variant="h6" paragraph>
+                                    Capturas de pantalla
+                                </Typography>
+                                <Box display="flex" flexWrap="wrap">
+                                    {
+                                        product.screenshots.map((s, i) => {
+                                            return <Box key={i} p={1}>
+                                                <img src={product.screenshots[i]} alt="Screenshot" className={classes.screenshot}/>
+                                            </Box>
+                                        })
+                                    }
+                                </Box>
+                            </div>
+                        }
+
                     </Box>
                 </div>
-                }
 
-                {product.screenshots && Array.isArray(product.screenshots) && product.screenshots.length > 0 &&
-                <div>
-                    <Typography variant="h6" paragraph>
-                        Capturas de pantalla
-                    </Typography>
-                    <Box display="flex" flexWrap="wrap">
-                        {
-                            product.screenshots.map((s, i) => {
-                                return <Box key={i} p={1}>
-                                    <img src={product.screenshots[i]} alt="Screenshot" className={classes.screenshot}/>
-                                </Box>
-                            })
-                        }
-                    </Box>
-                </div>
-                }
             </Grid>
         </Grid>
 
