@@ -1,5 +1,5 @@
 import React from 'react';
-import ResponsiveDrawer from "./ResposiveDrawer";
+import ResponsiveDrawer from "./ResposiveDrawerV2";
 //import MiniDrawer from "./MiniDrawer";
 
 const Home = () => {
